locally run wip webservice sqlite
Prompt:
Analyze how currently the EVALUATION_SQLITE_PATH is handled. is there env var that enables the game execution info to databse persistnace for teh evaluations entrypoints?

Analyze how would it be possible to extend persisting the game execution data into the sqlite database, when the game is running from the webservice.py entrypoint, and respect the parallel sessions.

I want to save the sqlite db file in a volume, that will be mounted in the container in a serverless environemnt with autoscaling enabled for the container.
Implement a lock path that is just a file with a 0/1 value next to the db file with a value inside that indicates whether any container instance in the cloud service (eg.via GCP cloud run) is already writing to that DB. If the value is set, prevent writing anything to the sqldb file.
My goal with this design is to enable saving the game execution data from one single container replica, and deny all subsequent replicas to write to the sqlite db file on the shared volume.